requirements.txt
pandas
requests
apify-client
python-dotenv
matplotlib 
seaborn 
folium
pyyaml